% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{as_json_string}
\alias{as_json_string}
\title{Return JSON string for response message}
\usage{
as_json_string(tagged)
}
\arguments{
\item{tagged}{baikalNLP tagger result}
}
\value{
returns JSON string
}
\description{
결과를 JSON 문자열로 변환.
}
\examples{
t <- tagger("결과를 문자열로 바꾼다.")
json <- as_json_string(t)
}
