% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{get_dict}
\alias{get_dict}
\title{Get Custom Dictionary}
\usage{
get_dict(tagged, name)
}
\arguments{
\item{tagged}{baikalNLP tagger result}

\item{name}{name of custom dictionary}
}
\value{
returns dict
}
\description{
사용자 사전 적용.
}
