% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{morphs}
\alias{morphs}
\title{Return array of Morphemes}
\usage{
morphs(m)
}
\arguments{
\item{message}{baikalNLP response message}
}
\value{
returns array of list for morphemes
}
\description{
형태소 분석 결과의 음절 배열 반환.
}
