% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{postag}
\alias{postag}
\title{Returns array of word, postag pairs}
\usage{
postag(m)
}
\arguments{
\item{message}{baikalNLP response message}
}
\value{
array of list result array of postags
}
\description{
결과를 (음절, 형태소 태그) 의 배열로 출력.
}
