% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{verbs}
\alias{verbs}
\title{Return array of Verbs}
\usage{
verbs(m)
}
\arguments{
\item{message}{baikalNLP response message}
}
\value{
returns array of list for verbs
}
\description{
형태소 분석 결과의 동사 배열 반환.
}
\examples{
> m <- tagger("결과를 문자열로 바꾼다.")
> verbs(m)
$verbs
[1] "바꾸"
}
