% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{morphs}
\alias{morphs}
\title{Returns array of Morphemes}
\usage{
morphs(m)
}
\arguments{
\item{message}{baikalNLP response message}
}
\value{
array of list result array of morphemes
}
\description{
형태소 분석 결과의 음절만 배열로 출력.
}
