% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{postag}
\alias{postag}
\title{Return array of (word, postag) pairs}
\usage{
postag(m)
}
\arguments{
\item{message}{baikalNLP response message}
}
\value{
returns array of list for (morpheme, postag)
}
\description{
결과에서 (음절, 형태소 태그)의 배열을 반환.
}
