% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{print_as_json}
\alias{print_as_json}
\title{Print JSON string for response message}
\usage{
print_as_json(message)
}
\arguments{
\item{message}{baikalNLP response message}
}
\value{
prints JSON string
}
\description{
결과를 JSON 문자열로 출력.
}
