% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{tagger}
\alias{tagger}
\title{Call baikalNLP server to read postag result message for the sentences}
\usage{
tagger(str, host = "nlp.baikal.ai:5656")
}
\arguments{
\item{str}{\itemize{
\item subject sentences splitted by newline(\\n)
}}

\item{host}{\itemize{
\item baikalNLP grpc server address
}}
}
\value{
returns response message of baikalNLP.AnalyzeSyntax
}
\description{
baikalNLP grpc 서버를 호출하여 입력 문장(들)의 분석 결과를 가져 온다.
}
\examples{
message <- tagger("결과를 문자열로 바꾼다.")
}
