% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baikalnlp.R
\name{tagger}
\alias{tagger}
\title{Call baikalNLP server to read postag result message for the sentences}
\usage{
tagger(text = "", server = "nlp.baikal.ai", port = 5656, local = FALSE)
}
\arguments{
\item{text}{string - subject sentences splitted by newlines(\\n)}

\item{server}{string - baikalNLP grpc server address}

\item{port}{number - baikalNLP grpc server port}

\item{local}{bool - use local protobuf files, if TRUE}
}
\value{
returns tagged object
}
\description{
baikalNLP grpc 서버를 호출하여 입력 문장(들)의 분석 결과를 가져 온다.
}
\examples{
tagged <- tagger("결과를 문자열로 바꾼다.")
}
